requirements.txt
asgiref==3.5.0
autopep8==1.6.0
bleach==4.1.0
click==8.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.1
django-environ==0.8.1
django-heroku==0.3.1
djlint==0.7.6
gunicorn==20.1.0
importlib-metadata==4.11.3
packaging==21.3
pathspec==0.9.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomlkit==0.9.2
tqdm==4.63.1
tzdata==2021.5
webencodings==0.5.1
whitenoise==6.0.0
zipp==3.7.0
